Cosmetics, resolve babenko-issues b6044e29b6bb543be4bcd074c38b1e79bbce0032

+++ yt/yt/server/lib/hive/logical_time_registry.h
@@ -18,7 +18,7 @@ namespace NYT::NHiveServer {
 struct TConsistentState
 {
     i64 SequenceNumber;
-    i64 SegmentId;
+    int SegmentId;
 };
 
 class TLogicalTimeRegistry
